player screen

+++ 2024-1C-3051-fossers/src/Events/Player.cs
@@ -13,12 +13,12 @@ public static class PlayerEvents
         EventManager.Instance.StartListening(Constants.Events.Player.RELOADING, cb);
     }
 
-    public static void TriggerHealthChanged(int newHealthValue)
+    public static void TriggerHealthChanged(float newHealthValue)
     {
         EventManager.Instance.TriggerEvent(Constants.Events.Player.HEALTH_CHANGED, newHealthValue);
     }
 
-    public static void SubscribeToHealthChanged(Action<int> cb)
+    public static void SubscribeToHealthChanged(Action<float> cb)
     {
         EventManager.Instance.StartListening(Constants.Events.Player.HEALTH_CHANGED, cb);
     }
